requirements.txt
kafka-python==2.0.2
msgpack==1.0.4
python-dotenv==0.20.0
